Add tests for service layer. Add allure module.

+++ src/main/java/com/brazhnyk/epam_finalproject_spring/service/implementation/EditionService.java
@@ -1,29 +1,19 @@
-package com.brazhnyk.epam_finalproject_spring.service;
+package com.brazhnyk.epam_finalproject_spring.service.implementation;
 
 import com.brazhnyk.epam_finalproject_spring.entity.Edition;
 import com.brazhnyk.epam_finalproject_spring.entity.Genre;
 import com.brazhnyk.epam_finalproject_spring.entity.User;
 import com.brazhnyk.epam_finalproject_spring.exception.AuthenticationError;
 import com.brazhnyk.epam_finalproject_spring.repository.EditionRepo;
-import com.brazhnyk.epam_finalproject_spring.util.InputValidator;
+import com.brazhnyk.epam_finalproject_spring.service.IEditionService;
 import com.brazhnyk.epam_finalproject_spring.util.PaginationPresetEngine;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.Locale;
-import java.util.Optional;
-
-import static com.brazhnyk.epam_finalproject_spring.util.PaginationPresetEngine.CURRENT_PAGE;
-import static com.brazhnyk.epam_finalproject_spring.util.PaginationPresetEngine.RECORDS_PER_PAGE;
-
 @Service
-public class EditionService {
+public class EditionService implements IEditionService {
 
     private final EditionRepo editionRepo;
 
@@ -32,6 +22,7 @@ public class EditionService {
         this.editionRepo = editionRepo;
     }
 
+    @Override
     public Page<Edition> findAll(String page, String records,Genre genre, String orderBy) {
 
         Pageable pageable = PaginationPresetEngine.definePageableByParam(page, records, orderBy);
@@ -43,40 +34,21 @@ public class EditionService {
         }
     }
 
+    @Override
     public void saveEdition(Edition edition) {
         if (edition != null) {
             editionRepo.save(edition);
         }
     }
 
-    public Edition updateEdition(Edition editionFromDb, Edition newEdition) {
-        BeanUtils.copyProperties(newEdition, editionFromDb, "id");
-
-        return editionRepo.save(newEdition);
+    @Override
+    public void deleteEdition(Edition edition) {
+        if (edition != null) {
+            editionRepo.delete(edition);
+        }
     }
 
-    public void delete(Edition edition) {
-        editionRepo.delete(edition);
-    }
-
-    public Optional<Edition> findById(Long editionId) {
-        return editionRepo.findById(editionId);
-    }
-
-    public Page<Edition> getEditionPage(String page, String records) {
-        Pageable pageable = PaginationPresetEngine.definePageableByParam(page, records, null);
-
-        return editionRepo.findAll(pageable);
-    }
-
-    /**
-     * Generates page with editions using provided filters.
-     * @param user - logged in user
-     * @param page - number of required page
-     * @param records - number of records per page
-     * @param genre - genreFilter
-     * @return page with editions
-     */
+    @Override
     public Page<Edition> findAllNotOrdered(User user, String page, String records, Genre genre, String orderBy) {
 
         Pageable pageable = PaginationPresetEngine.definePageableByParam(page, records, orderBy);
@@ -98,6 +70,7 @@ public class EditionService {
         }
     }
 
+    @Override
     public Page<Edition> findAllOrdered(User user, String page, String records,Genre genre, String orderBy) throws AuthenticationError {
 
         Pageable pageable = PaginationPresetEngine.definePageableByParam(page, records, orderBy);
